# src/modules/auth/auth.controller.ts
import { Body, Controller, Get, HttpCode, Post, Req, Res, UseGuards } from '@nestjs/common'; import { ApiBearerAuth, ApiTags } from '@nestjs/swagger'; import { Throttle } from '@nestjs/throttler'; import { Request, Response } from 'express'; import { CurrentUser } from '../../common/decorators/current-user.decorator'; import { AuthUser } from '../../common/types/auth-user'; import { AuthService, SessionResult } from './auth.service'; import { LoginDto } from './dto/login.dto'; import { JwtAuthGuard } from './jwt-auth.guard';
type CookieRequest = Request & { cookies?: Record<string, string> };
@ApiTags('Auth') @Controller('auth') export class AuthController { constructor(private readonly auth: AuthService) {} private send(result: SessionResult, response: Response): Omit<SessionResult, 'refreshToken'> { response.cookie('maryme_refresh', result.refreshToken, { httpOnly: true, secure: process.env.COOKIE_SECURE === 'true', sameSite: 'strict', path: '/api/v1/auth', maxAge: 30 * 86400000 }); const { refreshToken: _refreshToken, ...body } = result; return body; }
  @Post('login') @Throttle({default:{limit:5,ttl:60000}}) @HttpCode(200) async login(@Body() dto: LoginDto, @Req() req: Request, @Res({ passthrough: true }) res: Response) { return this.send(await this.auth.login(dto, req.ip, req.get('user-agent')), res); }
  @Post('refresh') @Throttle({default:{limit:20,ttl:60000}}) @HttpCode(200) async refresh(@Req() req: CookieRequest, @Res({ passthrough: true }) res: Response) { return this.send(await this.auth.refresh(req.cookies?.maryme_refresh, req.ip, req.get('user-agent')), res); }
  @Post('logout') @HttpCode(204) async logout(@Req() req: CookieRequest, @Res({ passthrough: true }) res: Response): Promise<void> { await this.auth.logout(req.cookies?.maryme_refresh); res.clearCookie('maryme_refresh', { path: '/api/v1/auth' }); }
  @Get('me') @UseGuards(JwtAuthGuard) @ApiBearerAuth() me(@CurrentUser() user: AuthUser) { return this.auth.me(user.sub); }
}