# src/modules/couples/couples.controller.ts
import { Body, Controller, Delete, Get, HttpCode, Param, Patch, Post, Query, UseGuards } from '@nestjs/common'; import { ApiBearerAuth, ApiTags } from '@nestjs/swagger'; import { UserRole, CoupleStatus } from '@prisma/client'; import { CurrentUser } from '../../common/decorators/current-user.decorator'; import { Roles } from '../../common/decorators/roles.decorator'; import { RolesGuard } from '../../common/guards/roles.guard'; import { AuthUser } from '../../common/types/auth-user'; import { JwtAuthGuard } from '../auth/jwt-auth.guard'; import { CouplesService } from './couples.service'; import { CoupleQueryDto, CreateCoupleDto, UpdateCoupleDto } from './dto/couple.dto';
@ApiTags('Couples') @ApiBearerAuth() @UseGuards(JwtAuthGuard,RolesGuard) @Controller('couples') export class CouplesController {constructor(private readonly service:CouplesService){} @Get() @Roles(UserRole.SUPER_ADMIN) list(@Query()q:CoupleQueryDto){return this.service.list(q)} @Post() @Roles(UserRole.SUPER_ADMIN) create(@Body()d:CreateCoupleDto,@CurrentUser()u:AuthUser){return this.service.create(d,u)} @Get('me') @Roles(UserRole.COUPLE) me(@CurrentUser()u:AuthUser){return this.service.get(u.coupleId!,u)} @Get(':id') get(@Param('id')id:string,@CurrentUser()u:AuthUser){return this.service.get(id,u)} @Patch(':id') update(@Param('id')id:string,@Body()d:UpdateCoupleDto,@CurrentUser()u:AuthUser){return this.service.update(id,d,u)} @Delete(':id') @Roles(UserRole.SUPER_ADMIN) @HttpCode(204) remove(@Param('id')id:string,@CurrentUser()u:AuthUser){return this.service.remove(id,u)} @Post(':id/authorize') @Roles(UserRole.SUPER_ADMIN) authorize(@Param('id')id:string,@CurrentUser()u:AuthUser){return this.service.status(id,CoupleStatus.AUTHORIZED,u)} @Post(':id/suspend') @Roles(UserRole.SUPER_ADMIN) suspend(@Param('id')id:string,@CurrentUser()u:AuthUser){return this.service.status(id,CoupleStatus.SUSPENDED,u)} }